Slider de rango funcionando a falta de estilos

+++ src/app/informe-filters/components/perdidas-filter/perdidas-filter.component.ts
@@ -11,25 +11,29 @@ import { Options } from '@angular-slider/ngx-slider';
 @Component({
   selector: 'app-perdidas-filter',
   templateUrl: './perdidas-filter.component.html',
-  styleUrls: ['./perdidas-filter.component.css'],
+  styleUrls: ['./perdidas-filter.component.scss'],
 })
 export class PerdidasFilterComponent implements OnInit {
   rangoMinPerdidas: number;
+  rangoMaxPerdidas: number;
   filtroPerdidas: PerdidasFilter;
-  options: Options = { floor: 0, ceil: 200 };
+  options: Options = { floor: 0, ceil: 100 };
 
-  constructor(private filterService: FilterService) {}
+  constructor(private filterService: FilterService) {
+    this.rangoMinPerdidas = 0;
+    this.rangoMaxPerdidas = 100;
+  }
 
   ngOnInit(): void {}
 
-  onInputFiltroPerdidas(event: MatSliderChange) {
+  /* onInputFiltroPerdidas(event: MatSliderChange) {
     this.rangoMinPerdidas = event.value;
-  }
+  } */
 
   onChangeFiltroPerdidas() {
-    this.filtroPerdidas = new PerdidasFilter('perdidas', this.rangoMinPerdidas, 100);
+    this.filtroPerdidas = new PerdidasFilter('perdidas', this.rangoMinPerdidas, this.rangoMaxPerdidas);
 
-    if (this.rangoMinPerdidas === 0) {
+    if (this.rangoMinPerdidas === 0 && this.rangoMaxPerdidas === 100) {
       // si se selecciona el mínimo desactivamos el filtro ...
       this.filterService.deleteFilter(this.filtroPerdidas);
     } else {
